chore: move honor
