+++ api/src/pdc/services/honor/actions/SaveAction.ts
@@ -1,13 +1,9 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 
-import { alias } from "@/shared/honor/save.schema.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/honor/save.contract.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "../contracts/save.contract.ts";
+import { alias } from "../contracts/save.schema.ts";
 import { HonorRepositoryInterfaceResolver } from "../providers/HonorRepositoryProvider.ts";
 
 @handler({
